+++ packages/agent/CHANGELOG.md
@@ -14,6 +14,10 @@
 
 ## [2026.5.29-4] - 2026-05-29
 
+## [0.78.0] - 2026-05-29
+
+## [0.77.0] - 2026-05-28
+
 ### Breaking Changes
 
 - Renamed agent harness `model_select` and `thinking_level_select` events to `model_update` and `thinking_level_update`.